backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cnocr==2.3.2.2
cnstd==1.2.6.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
evaluate==0.4.6
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openai==2.17.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
optimum==1.16.2
optimum-onnx==0.1.0
packaging==26.0
pandas==2.3.3
pillow==10.2.0
pix2text==1.1.4
platformdirs==4.5.1
polars==1.37.1
polars-runtime-32==1.37.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyMuPDF==1.26.7
pyparsing==3.3.2
pyspellchecker==0.8.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.6.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.51.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
thop==0.1.1.post2209072238
tokenizers==0.22.2
torch==2.2.0
torchmetrics==1.8.2
torchvision==0.17.0
tqdm==4.67.2
transformers==4.57.6
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.11
ultralytics-thop==2.0.18
Unidecode==1.4.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
wandb==0.24.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
